requirements.txt
# 基础依赖
pandas==2.1.0
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# 数据处理
akshare==1.11.80
efinance==0.5.6
tushare==1.4.5

# AI相关
openai==1.3.0
anthropic==0.8.1
tiktoken==0.5.1

# 微调相关
transformers==4.36.0
torch==2.1.0
datasets==2.15.0
peft==0.7.1
accelerate==0.25.0
bitsandbytes==0.41.3
sentencepiece==0.1.99

# 工具
tqdm==4.66.1
python-dotenv==1.0.0
loguru==0.7.2

